Add `SizeConstraints` to `Axis` and its subclasses

+++ vico/core/src/main/java/pl/patrykgoworowski/vico/core/axis/AxisRenderer.kt
@@ -83,5 +83,5 @@ public interface AxisRenderer<Position : AxisPosition> : BoundsAware, ChartInset
 
     public fun setRestrictedBounds(vararg bounds: RectF?)
 
-    public fun getDesiredHeight(context: MeasureContext): Int
+    public fun getDesiredHeight(context: MeasureContext): Float
 }
